requirements.etl.txt
gunicorn==21.2.0
python-dotenv==1.0.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
requests==2.32.3
pandas==2.2.3
